Move ApiConsumer to 'components'.
