+++ application/common/components/ApiConsumer.php
@@ -1,5 +1,5 @@
 <?php
-namespace Sil\Idp\IdSync\common\models;
+namespace Sil\Idp\IdSync\common\components;
 
 use yii\base\Component;
 use yii\web\IdentityInterface;
